+++ mitra-cli/src/main.rs
@@ -31,7 +31,7 @@ async fn main() {
             match subcmd {
                 SubCommand::GenerateInviteCode(cmd) => cmd.execute(db_client).await.unwrap(),
                 SubCommand::ListInviteCodes(cmd) => cmd.execute(db_client).await.unwrap(),
-                SubCommand::CreateUser(cmd) => cmd.execute(db_client).await.unwrap(),
+                SubCommand::CreateUser(cmd) => cmd.execute(&config, db_client).await.unwrap(),
                 SubCommand::ListUsers(cmd) => cmd.execute(db_client).await.unwrap(),
                 SubCommand::AddEd25519Key(cmd) => cmd.execute(db_client).await.unwrap(),
                 SubCommand::SetPassword(cmd) => cmd.execute(db_client).await.unwrap(),